+++ lib/src/vector/common/vector_data_store.dart
@@ -1,3 +1,0 @@
-abstract class VectorDataStore<E, S extends List<E>> {
-  S data;
-}