+++ test/test_all.dart
@@ -1,6 +1,6 @@
-import 'float32x4_matrix_test.dart' as float32x4_matrix_test;
+import 'float32_matrix_test.dart' as float32x4_matrix_test;
 import 'matrix_iterator_test.dart' as matrix_iterator_test;
-import 'float32x4_vector_test.dart' as float32x4_vector_test;
+import 'float32_vector_test.dart' as float32x4_vector_test;
 
 void main() {
   float32x4_matrix_test.main();