+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 10.0.0
+- `Vector`:
+    - removed possibility to mutate an instance of the Vector:
+        - `isMutable` flag removed
+        - `[]=` operator removed
+- `Matrix`:
+    - removed possibility to mutate an instance of the Matrix:
+        - `setColumn` method removed
+    - `insertColumns` method added
+    - `Matrix.from` renamed to `Matrix.fromList`
+    - `Matrix.fromFlattened` renamed to `Matrix.fromFlattenedList` 
+    - benchmarks added for checking matrix initialization performance
+- Performance enhancements
+- Dart 2.3.0 supported
+
 ## 9.0.0
 - `Vector`:
     - `Distance` enum added